Added a tile engine, basic duck AI and smart Camera
I have created a tile engine which will be used to create the maze.
Currently there is only one type of tile but more will be added. And the
tile isn't textured. Duck moves forward along the base of the maze. No
physics implementation. If the player is closer than screenwidth/2 or
screenheight/2 to an edge the camera will stay at that
location, preventing the user from seeing outside the map but still
allowing the player to walk up to the edges.