Implemented collision detection and momentum
Collision detection now works as intended. Changed the way movement was
being done so when the duck is falling it takes into account the
movement velocity and vector. Also allowed the user to rotate the camera
with the up and down keys which subsequently changes the gravity vector
as well. However the method of rotating the camera is still broken
(zooms out).