Physics merged with all other additions
Physics work flawlessly on left, right and down sides. Haven't had a
chance to test it with the gravity flipping ability or PS Vita
accelerometer support (haven't coded in bird rotation). All this has
been merged with the rest of the new features. I have left the bouncing
in for testing purpose.