HUD, additional camera mode, collision detection changes
Implemented a HUD that includes a working score and timer. Timer
currently works in seconds without flaw. Score is incremented via coin
and gem pickups, with coins 10 and gems 20. HUD correctly rotates when
the user rotates the screen. Added an additional camera mode which can
be toggled between during the game. Collision detection is still being
worked on.