[+] Added movement controls for the simulator
Added the ability to move without having a motion enabled device. This
will allow for extra testing and designing whilst on the simulator. To
move use the arrow keys. KNOWN BUG: When falling parallel to the X axis
it's common to clip into tiles if you go too close to the sides.