[+] Third Level, Select Screen Back Button, Spawns
Added a third level which demonstrates lasers. Added functionality to
the back button on the level select screen although there's a slight
issue with progressing again once you've gone back but will fix soon.
Added the ability to set spawns via the XML Level's. This was partly
there already but made sure it works now. Fixed a issue with collisions
when falling to the right when you would go through the walls. It is now
pretty had to mess up the collisions.