[+] Animated Laser, Updated Black Hole, Collisions
Added a new laser which is animated. Working on getting as close to
pixel perfect laser collisions as possible. Black hole is now actually a
threat and will suck you in to it as intended. If you get fully sucked
in to the black hole you will die. Updated the collisions so that if you
collide with a side wall when falling from left to right (and vice
versa) you won't get a juttery effect and will get a smooth rebound.