[+] End level buttons fully work + related fixes
You can now return to the level select screen once you have completed a
level. The level complete screen and the game over screen have both been
given a new function which essentially reorders them in the Z axis,
bringing them to the front of the screen. Previously the were shown
behind the maze and couldn't be seen.