[+] Added 21 tester levels, play and performance testing
21 basic levels have been added, giving us the full 27 levels. The
purpose of this was to test progression through levels and to ensure you
could move throughout the non-gameplay part of game. I foresaw a
potential problem with memory on the Vita so I removed the loading
screen textures from memory during the levels and brought them back when
the user is on the menus.