[+] Better memory management + more backgrounds
I have changed the way the loading screen textures are being used.
Previously there were 3 iterations of each of the 27 loading screens in
memory. I have changed that to be just 1 iteration of all of them whilst
in the level select screen and only the required loading screen whilst
in the game. I have added 4 in-game backgrounds and will need to get
some feedback to check if they are in the right style. I have
temporarily removed the options and hiscores button off the title
screen. I have also fixed the issue where entities were remaining on the
screen after the player has progressed to the next level.